Changed render method for views and layouts.
The previous render method allowed you to render a template
directly into a layout. This however only allowed one template
to be used per layout. The new method instead stores the template
results as a view variable, which the layout can reference.

Also expanded documentation on views.

Added a has() method to the view to check if a variable is set.